Multi-platform improvement
Created vck5000 and v70 files and moved related APIs to their platform files.
Fixed a few bugs from PR #156
