+++ vmr/src/vmc/platforms/v70.c
@@ -1,0 +1,8 @@
+/******************************************************************************
+* Copyright (C) 2022 Xilinx, Inc.  All rights reserved.
+* SPDX-License-Identifier: MIT
+*******************************************************************************/
+
+#include "v70.h"
+
+
