+++ vmr/src/vmc/platforms/v70.h
@@ -1,0 +1,12 @@
+/******************************************************************************
+* Copyright (C) 2022 Xilinx, Inc.  All rights reserved.
+* SPDX-License-Identifier: MIT
+*******************************************************************************/
+
+#ifndef INC_PLATFORMS_V70_H_
+#define INC_PLATFORMS_V70_H_
+
+#include "../vmc_asdm.h"
+
+
+#endif
